requirements.txt
django==3.2
djangorestframework
mysql-client
python-dotenv
pymysql